--- a/models/gold/gov/gov__fact_stake_accounts.sql
+++ b/models/gold/gov/gov__fact_stake_accounts.sql
@@ -82,3 +82,4 @@ SELECT
   '2000-01-01' AS modified_timestamp
 FROM
   {{ ref('silver__historical_stake_account') }}
+{% endif %}