--- a/models/gold/gov/gov__fact_validators.sql
+++ b/models/gold/gov/gov__fact_validators.sql
@@ -1,8 +1,23 @@
 {{ config(
-  materialized = 'view',
-  meta ={ 'database_tags':{ 'table':{ 'PURPOSE': 'VALIDATOR' }} },
-  tags = ['scheduled_non_core','exclude_change_tracking']
+    materialized = 'incremental',
+    meta ={ 'database_tags':{ 'table':{ 'PURPOSE': 'VALIDATOR' }}},
+    unique_key = ['fact_validators_id'],
+    cluster_by = ['epoch','epoch_active'],
+    merge_exclude_columns = ["inserted_timestamp"],
+    post_hook = enable_search_optimization('{{this.schema}}','{{this.identifier}}','ON EQUALITY(node_pubkey, vote_pubkey)'),
+    tags = ['scheduled_non_core','exclude_change_tracking']
 ) }}
+
+{% if execute %}
+    {% if is_incremental() %}
+        {% set query %}
+            SELECT MAX(modified_timestamp) AS max_modified_timestamp
+            FROM {{ this }}
+        {% endset %}
+
+        {% set max_modified_timestamp = run_query(query).columns[0].values()[0] %}
+    {% endif %}
+{% endif %}
 
 SELECT
   epoch_recorded :: INT AS epoch,
@@ -42,6 +57,11 @@ SELECT
   ) AS modified_timestamp
 FROM
   {{ ref('silver__snapshot_validators_app_data') }}
+{% if is_incremental() %}
+WHERE
+    modified_timestamp >= '{{ max_modified_timestamp }}'
+{% endif %}
+{% if not is_incremental() %}
 UNION ALL
 SELECT
   epoch_recorded :: INT AS epoch,
@@ -72,3 +92,4 @@ SELECT
   '2000-01-01' AS modified_timestamp
 FROM
   {{ ref('silver__historical_validator_app_data') }}
+{% endif %}