--- a/models/silver/metadata/helius/silver__helius_nft_metadata.sql
+++ b/models/silver/metadata/helius/silver__helius_nft_metadata.sql
@@ -3,6 +3,7 @@
     unique_key = "mint",
     incremental_strategy = 'delete+insert',
     cluster_by = ['_inserted_timestamp::DATE'],
+    post_hook = enable_search_optimization('{{this.schema}}', '{{this.identifier}}', 'ON EQUALITY(mint,nft_name,nft_collection_id)'),
     tags = ['nft_api']
 ) }}