--- a/models/silver/nfts/silver__nft_burn_actions.sql
+++ b/models/silver/nfts/silver__nft_burn_actions.sql
@@ -3,6 +3,7 @@
     unique_key = "CONCAT_WS('-', tx_id, index, inner_index, mint)",
     incremental_strategy = 'delete+insert',
     cluster_by = ['block_timestamp::DATE'],
+    post_hook = enable_search_optimization('{{this.schema}}', '{{this.identifier}}', 'ON EQUALITY(tx_id,mint,burn_authority)'),
     tags = ['scheduled_non_core']
 ) }}