--- a/models/silver/nfts/silver__nft_mints.sql
+++ b/models/silver/nfts/silver__nft_mints.sql
@@ -3,6 +3,7 @@
     unique_key = "CONCAT_WS('-', initialization_tx_id, mint, purchaser, mint_currency)",
     incremental_strategy = 'delete+insert',
     cluster_by = ['block_timestamp::DATE'],
+    post_hook = enable_search_optimization('{{this.schema}}', '{{this.identifier}}', 'ON EQUALITY(tx_id,purchaser,mint)'),
     tags = ['scheduled_non_core']
 ) }}