--- a/models/silver/nfts/silver__nft_sales_hadeswap.sql
+++ b/models/silver/nfts/silver__nft_sales_hadeswap.sql
@@ -2,7 +2,7 @@
   materialized = 'incremental',
   unique_key = "CONCAT_WS('-', tx_id, mint)",
   incremental_strategy = 'delete+insert',
-  cluster_by = ['block_timestamp::DATE'],
+  cluster_by = ['block_timestamp::DATE','modified_timestamp::DATE'],
   tags = ['scheduled_non_core']
 ) }}