--- a/models/silver/nfts/silver__nft_sales_solsniper_cnft.sql
+++ b/models/silver/nfts/silver__nft_sales_solsniper_cnft.sql
@@ -3,6 +3,7 @@
     unique_key = ['nft_sales_solsniper_cnft_id'],
     incremental_strategy = 'delete+insert',
     merge_exclude_columns = ["inserted_timestamp"],
+    cluster_by = ['block_timestamp::DATE','modified_timestamp::DATE'],
     tags = ['scheduled_non_core']
 ) }}