--- a/models/silver/validator/silver__snapshot_vote_accounts.sql
+++ b/models/silver/validator/silver__snapshot_vote_accounts.sql
@@ -2,7 +2,7 @@
     materialized = 'incremental',
     unique_key = "CONCAT_WS('-', epoch_recorded, vote_pubkey)",
     incremental_strategy = 'delete+insert',
-    cluster_by = ['_inserted_timestamp::DATE'],
+    cluster_by = ['modified_timestamp::DATE'],
     tags = ['validator']
 ) }}