Search optimization and clustering for NFT and GOV schema

1. For larger/complex core views with union/join logic, convert to table and update silver clustering
- add logic to select from legacy views only during FR
- includes: `gov.fact_stake_accounts`, `gov.fact_validators`, `gov.fact_vote_accounts`, `nft.fact_nft_sales`)

3. For views with simpler logic, add SO and clustering to silver references

FR successful for core tables (on 2xl):
```
-- 23:47:48  2 of 5 OK created sql incremental model gov.fact_validators .................... [SUCCESS 1 in 23.25s]
-- 23:47:53  3 of 5 OK created sql incremental model gov.fact_vote_accounts ................. [SUCCESS 1 in 28.33s]
-- 23:48:27  4 of 5 OK created sql incremental model nft.fact_nft_sales ..................... [SUCCESS 1 in 61.76s]
-- 23:48:37  1 of 5 OK created sql incremental model gov.fact_stake_accounts ................ [SUCCESS 1 in 71.95s]
```
Incremental runs for core tables (on medium):
```
02:59:22  3 of 5 OK created sql incremental model gov.fact_vote_accounts ................. [SUCCESS 6300 in 7.92s]
02:59:23  4 of 5 OK created sql incremental model nft.fact_nft_sales ..................... [SUCCESS 266 in 8.62s]
02:59:24  2 of 5 OK created sql incremental model gov.fact_validators .................... [SUCCESS 1512 in 9.75s]
02:59:35  1 of 5 OK created sql incremental model gov.fact_stake_accounts ................ [SUCCESS 1442330 in 20.77s]
```